+++ RELEASE_NOTES.md
@@ -5,6 +5,15 @@ StoreKitPlus will use semver after 1.0.
 Until then, deprecated features may be removed in any minor version, and breaking changes introduced.
 
 
+
+## 0.7
+
+### ✨ Features
+
+* `StoreService` now lets you pass in optional purchase options.
+
+
+
 ## 0.6
 
 This version removes deprecated code and adjusts some service terminology.